+++ lib/src/wip/language_syntax/golang_syntax.dart
@@ -11,8 +11,7 @@ Map<int, String> findGolangErrors(String text) {
     }
 
     if (lines[i].startsWith(RegExp("\\s*/\\*"))) {
-      while (
-          (!lines[i].contains(RegExp("\\*/\\s*"))) && (i < lines.length - 1)) {
+      while ((!lines[i].contains(RegExp("\\*/\\s*"))) && (i < lines.length - 1)) {
         i++;
       }
     } else if (lines[i].startsWith(RegExp(".*`"))) {
@@ -34,9 +33,8 @@ Map<int, String> findGolangErrors(String text) {
         lines[i] = lines[i];
       }
       commandFor += lines[i];
-      if (commandFor.contains(RegExp("for.*:=")) &&
-          commandFor.contains(RegExp("for.*[\+-][\+-]"))) {
-        if (!commandFor.contains(RegExp("for.*:=.*;.*;.*[\+-][\+-]"))) {
+      if (commandFor.contains(RegExp("for.*:=")) && commandFor.contains(RegExp("for.*[+-][+-]"))) {
+        if (!commandFor.contains(RegExp("for.*:=.*;.*;.*[+-][+-]"))) {
           errors.addAll({(i + 1): "Incorrect for statement"});
         }
       }