+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.3.0
+
+* **BREAKING:** Deleted `CodeController.stringMap`.
+* **BREAKING:** Deleted `theme` from `CodeController`, (Issue [172](https://github.com/akvelon/flutter-code-editor/issues/172)).
+
+## 0.2.24
+
+* Hide the suggestion box when `CodeField` is disposed (Issue [241](https://github.com/akvelon/flutter-code-editor/issues/241)).
+* Fix issues with Flutter 3.10.
+
 ## 0.2.23
 
 * Added `CodeController.readOnly`.